lead-generation-agent/requirements.txt
fastapi==0.115.11
uvicorn==0.28.0
python-dotenv==1.0.1
pydantic==2.6.4
langgraph==0.0.40
groq==0.5.4
tavily-python==0.3.1
asyncio==3.4.3
websockets==12.0
python-multipart==0.0.7
markdown2==2.5.0
pymongo==4.6.2
